pyproject.toml
[project]
name = "authsocials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth[with-social]>=7.2.0",
    "django>=6.0.4",
    "django-allauth>=65.16.1",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
]
prod = [
    "gunicorn>=25.3.0",
]
